.WebcamServer/requirements.txt
mediapipe
opencv-python
asyncio
websockets
python-dotenv
python-decouple